Fix T38661: make number button increment/decrement areas smaller.
Previously 1/3 of the button was used to decrement, 1/3 to edit and 1/3 to
increment. However with the number text now right aligned this meant that
the increment area would overlap the number text, which is confusing. So it
was made to smaller to only cover the arrows.

It's not as easy to click but I don't know of a better solution with right
aligned number text.